[WFCORE-7143] Consolidate HostControllerBootstrap and EmbeddedHostControllerBootstrap
